Move goodreads functions to separate file
