+++ audible.py
@@ -3,7 +3,6 @@ import os
 import pathlib
 import re
 from tempfile import NamedTemporaryFile
-from typing import Dict
 import yaml
 
 from beets import importer, util
@@ -13,7 +12,8 @@ from beets.plugins import BeetsPlugin, get_distance
 import mediafile
 from natsort import os_sorted
 
-from .api import get_book_info, make_request, search_audible, search_goodreads
+from .api import get_book_info, make_request, search_audible
+from .goodreads import get_original_date
 
 class Audible(BeetsPlugin):
     data_source = 'Audible'
@@ -365,7 +365,7 @@ class Audible(BeetsPlugin):
         original_day=day
 
         if self.config['goodreads_apikey']:
-            original_date = self.get_original_date(asin, authors, title)
+            original_date = get_original_date(self, asin, authors, title)
             if original_date.get("year") is not None:
                 original_year=original_date.get("year")
                 original_month=original_date.get("month")
@@ -379,62 +379,6 @@ class Audible(BeetsPlugin):
             is_chapter_data_accurate=is_chapter_data_accurate,
             language=book.language, label=book.publisher, **common_attributes
         )
-
-    def get_original_date(self, asin: str, authors: str, title: str) -> Dict:
-        api_key = self.config['goodreads_apikey']
-        goodreads_response = search_goodreads(api_key, asin)
-        totalresults = self.goodreads_get_total_result(goodreads_response)
-
-        if totalresults == 0:
-            #search with author and title
-            #set asin to none to determine that we are querying on author/title
-            self._log.debug(f"search Goodreads again based on author/title.")
-            goodreads_response = search_goodreads(api_key, f"{authors} {title}")
-            totalresults = self.goodreads_get_total_result(goodreads_response)
-
-        self._log.debug(f"{totalresults} results found")
-        if totalresults >= 1:
-            work = self.goodreads_get_best_match(goodreads_response, authors, title)
-            original_date = self.parse_original_date(work)
-        else:
-            original_date = {}
-        
-        return original_date
-
-    def goodreads_get_best_match(self, response: Dict, author: str, title: str) -> Dict:
-        #returns best matching work from results
-        author_cleaned = author.replace(" ","")
-        #get all works
-        for work in response.findall("./search/results/work"):
-            best_book = work.find('best_book')
-            
-            #remove anything after parenthesis, this is where GR puts series and other non title info
-            gr_title = best_book.find('title').text
-            gr_title_cleaned = gr_title.split('(')[0].strip()
-
-            #remove all spaces from author name. Audible can have names like James S. A. Corey, GR might have James S.A. Corey
-            gr_author = best_book.find('author/name').text
-            gr_author_cleaned = gr_author.replace(" ","").strip()
-
-            #confirm author and titles
-            if (author_cleaned == gr_author_cleaned and title == gr_title_cleaned):
-                self._log.debug(f"Goodreads match found #{best_book.find('id').text} - {gr_author} {gr_title}")
-                return work
-
-    def goodreads_get_total_result(self, response: Dict) -> int:
-        return int(response.findtext("./search/total-results"))
-
-    def parse_original_date(self, work: Dict) -> Dict:
-        original_date = {}
-        if work is not None:
-            year = work.findtext("original_publication_year")
-            original_date["year"] = int(year) if year.isdigit() else None
-            month = work.findtext("original_publication_month")
-            original_date["month"] = int(month) if month.isdigit() else None
-            day = work.findtext("original_publication_day")
-            original_date["day"] = int(day) if day.isdigit() else None
-        
-        return original_date
     
     def on_write(self, item, path, tags):
         # Strip unwanted tags that Beets automatically adds
